+++ src/SwaggerProvider.DesignTime/v3/OperationCompiler.fs
@@ -2,14 +2,14 @@ namespace SwaggerProvider.Internal.v3.Compilers
 
 open System
 open System.Net.Http
-open System.Threading.Tasks
+open System.Text.Json
 open System.Text.RegularExpressions
 
 open Microsoft.FSharp.Quotations
 open Microsoft.FSharp.Quotations.ExprShape
+open Microsoft.OpenApi.Models
 open ProviderImplementation.ProvidedTypes
 open FSharp.Data.Runtime.NameUtils
-open Microsoft.OpenApi.Models
 
 open SwaggerProvider.Internal
 open Swagger
@@ -51,6 +51,11 @@ type OperationCompiler (schema:OpenApiDocument, defCompiler:DefinitionCompiler, 
 
         let unambiguousName (par: OpenApiParameter) =
             sprintf "%sIn%A" par.Name par.In
+        let openApiParameters =
+            seq {
+                yield! pathItem.Parameters
+                yield! operation.Parameters
+            } |> Seq.toList
         let parameters =
             /// handles deduping Swagger parameter names if the same parameter name
             /// appears in multiple locations in a given operation definition.
@@ -105,8 +110,7 @@ type OperationCompiler (schema:OpenApiDocument, defCompiler:DefinitionCompiler, 
 
             let required, optional =
                 seq {
-                    yield! pathItem.Parameters
-                    yield! operation.Parameters
+                    yield! openApiParameters
                     if bodyParam.IsSome
                     then yield bodyParam.Value
                 }
@@ -191,7 +195,7 @@ type OperationCompiler (schema:OpenApiDocument, defCompiler:DefinitionCompiler, 
                 |> List.choose (function
                     | ShapeVar sVar as expr ->
                         let param =
-                            operation.Parameters
+                            openApiParameters
                             |> Seq.tryFind (fun x ->
                                 // pain point: we have to make sure that the set of names we search for here are the same as the set of names generated when we make `parameters` above
                                 let baseName = niceCamelName x.Name
@@ -382,17 +386,41 @@ type OperationCompiler (schema:OpenApiDocument, defCompiler:DefinitionCompiler, 
 
             [
                 ProvidedConstructor(
-                    [ProvidedParameter("httpClient", typeof<HttpClient>)],
+                    [ProvidedParameter("httpClient", typeof<HttpClient>);
+                     ProvidedParameter("options", typeof<JsonSerializerOptions>)],
                     invokeCode = (fun args ->
                         match args with
                         | [] -> failwith "Generated constructors should always pass the instance as the first argument!"
                         | _ -> <@@ () @@>),
                     BaseConstructorCall = fun args -> (baseCtor, args))
-                ProvidedConstructor([],
+                ProvidedConstructor(
+                    [ProvidedParameter("httpClient", typeof<HttpClient>)],
+                    invokeCode = (fun args ->
+                        match args with
+                        | [] -> failwith "Generated constructors should always pass the instance as the first argument!"
+                        | _ -> <@@ () @@>),
+                    BaseConstructorCall = fun args ->
+                        let args' = args @ [ <@@ null @@> ]
+                        (baseCtor, args'))
+                ProvidedConstructor(
+                    [ProvidedParameter("options", typeof<JsonSerializerOptions>)],
                     invokeCode = (fun args -> <@@ () @@>),
                     BaseConstructorCall = fun args ->
-                        let httpClient = <@ RuntimeHelpers.getDefaultHttpClient defaultHost @>
-                        let args' = args @ [httpClient]
+                        let httpClient = <@ RuntimeHelpers.getDefaultHttpClient defaultHost @> :> Expr
+                        let args' =
+                            match args with
+                            | [instance; options] -> [instance; httpClient; options]
+                            | _ -> failwithf "unexpected arguments received %A" args
+                        (baseCtor, args'))
+                ProvidedConstructor(
+                    [],
+                    invokeCode = (fun args -> <@@ () @@>),
+                    BaseConstructorCall = fun args ->
+                        let httpClient = <@ RuntimeHelpers.getDefaultHttpClient defaultHost @> :> Expr
+                        let args' =
+                            match args with
+                            | [instance] -> [instance; httpClient; <@@ null @@>]
+                            | _ -> failwithf "unexpected arguments received %A" args
                         (baseCtor, args'))
             ] |> ty.AddMembers